Added support for docker-compose
Closes #25.
Now we can spin up multi-container docker environments, given a valid
docker compose yaml file.

Since ShellUtility was needed, jIcing was added as a jar dependecy to
this project instead of a git submodule. This was done as, if we were
to add it as a git submodule, then the maven packaging would have to
be changed to "pom", in which case we won't be able to execute the
unit tests via maven.

As of now there is only support for docker compose up and down.
Support for remaining compose commands can be added as needed later.